refactor: extract PageProxy from workflow.rb (v0.15 WS-7 PR 16)
